move Android files
