+++ android/CMakeLists.txt
@@ -24,18 +24,20 @@ add_library(
         ${PACKAGE_NAME}
         SHARED
         ../cpp/JSITypedArray.cpp
-        src/main/cpp/FrameHostObject.cpp
-        src/main/cpp/FrameProcessorPluginHostObject.cpp
-        src/main/cpp/JSIJNIConversion.cpp
         src/main/cpp/VisionCamera.cpp
-        src/main/cpp/VisionCameraProxy.cpp
+        # Frame Processor
+        src/main/cpp/frameprocessor/FrameHostObject.cpp
+        src/main/cpp/frameprocessor/FrameProcessorPluginHostObject.cpp
+        src/main/cpp/frameprocessor/JSIJNIConversion.cpp
+        src/main/cpp/frameprocessor/VisionCameraProxy.cpp
+        src/main/cpp/frameprocessor/java-bindings/JFrame.cpp
+        src/main/cpp/frameprocessor/java-bindings/JFrameProcessor.cpp
+        src/main/cpp/frameprocessor/java-bindings/JFrameProcessorPlugin.cpp
+        src/main/cpp/frameprocessor/java-bindings/JHashMap.cpp
+        src/main/cpp/frameprocessor/java-bindings/JVisionCameraProxy.cpp
+        src/main/cpp/frameprocessor/java-bindings/JVisionCameraScheduler.cpp
+        # Skia Frame Processor
         src/main/cpp/skia/SkiaRenderer.cpp
-        src/main/cpp/java-bindings/JFrame.cpp
-        src/main/cpp/java-bindings/JFrameProcessor.cpp
-        src/main/cpp/java-bindings/JFrameProcessorPlugin.cpp
-        src/main/cpp/java-bindings/JHashMap.cpp
-        src/main/cpp/java-bindings/JVisionCameraProxy.cpp
-        src/main/cpp/java-bindings/JVisionCameraScheduler.cpp
 )
 
 # Header Search Paths (includes)
