+++ android/src/main/cpp/frameprocessor/VisionCameraProxy.cpp
