+++ android/src/main/cpp/frameprocessor/VisionCameraProxy.h
