+++ android/src/main/cpp/frameprocessor/java-bindings/JFrameProcessorPlugin.cpp
