+++ android/src/main/cpp/frameprocessor/java-bindings/JVisionCameraProxy.h
