+++ android/src/main/cpp/frameprocessor/java-bindings/JVisionCameraScheduler.h
